Mayor refactor of fetch.c
This adds a char** to all the data collecting functions, which they can
use to store a pointer to an allocated buffer, which is then freed after
executinon. This we there is a smaller dependency on statically sized
buffers.
It also adds a second way to determine the shell of the user (via passwd)
changes exit(1) to exit(EXIT_FAILURE) removes scanf from distro.